gpp-woo-tiles component
